A digital clock using an Adafruit Neopixel Ring
The purpose of this project is to make a digital clock using an
Adafruit Neopixel 12-pixel Ring, an Adafruit RTC DS1307, an analog
potentiometer, a piezo speaker, a switch, and a button. These parts
will work together to make a digital clock in the shape of a ring,
whose alarm can be set using the potentiometer, switch, and button. The
Real Time Clock (RTC) will be synced with the users computer upon
upload of the program to the Arduino. The Neopixel ring will then
display the time given by the RTC on its pixels, with “hands” (simply
pixels that are on) for seconds, minutes, and hours. The alarm will be
set based on the position of the potentiometer when the button is
pressed.